replication: Add script for prune dead replicas
Introduce box.replication.get_alive_replicas and
box.replication.prune_dead_replicas methods.

The first method returns a table of replicas which are alive
within the specified time period.

The second method deletes all replicas which are not in table of alives
from box.space._cluster.

The alive replica in period is the one which has upstream.status or
downstream.status not equal to "stopped" or "disconnected" or the one
that was added during specified period.

Add test-run filter to some tests in engine package.

Closes #3110