update: rework updates to improve code extendibility
Before the patch update is implemented as a set of operations
applicable for arrays only. It was ok until field names and
JSON paths appearance, because tuple is array on the top level.

But now there are two reasons to allow more complex updates of
tuple field internals via JSON paths so far:
* tuple field access by JSON path is allowed so for consistency
  JSON paths should be allowed in updates as well;
* Tarantool is going to support documents in storage so JSON
  path updates is one more step forward.

The patch prepares current update code to be extended by updates
of map, so named 'bar' updates, and multilevel updates. No
functional changes.

Needed for #1261