update: enable JSON bar updates
Bar update is the update by JSON of a tuple internal object with
no other updates along the path. It is an optimization to do not
store each part of the path and link to it in memory. Bar stores
just string with JSON and pointer to the MessagePack object to
update.

So far bar can not have another updates along the path, this
patch introduces only non-intersected JSON path updates (that
is actually a most common case, so bar is really useful
optimization).

Part of #1261