Update lua space cache just after creation
The lua space cache (box.space.*) should be valid just after space is
created because space is ready to accept new records or does not exists
before wal would be written. So invoke a space create/drop trigger after
a space is changed and recall it in a case of rollback.

Relates: 3159