box: introduce box.ctl.promote
Replicaset master promotion is a procedure of atomic making one
slave be a new master, and an old master be a slave in a fullmesh
master-slave replicaset.

The promotion follows the protocol described in details in the
corresponding RFC. Shortly, the protocol collects a quorum of
instances who approves the promotion, syncs data with the old
master and demotes it.

The protocol is intended to work with a single master cluster and
with at least 50% + 1 quorum mandatory including an old master.
It is tolerant to messages reordering from different sources, to
errors like timeouts, restarts of any promotion participant. Also
the promote protocol supports promotion in a completely
read-only cluster. It is useful, for example, when after one of
rare cases of a promotion fail the cluster is left in a read-only
state with one already demoted master. Then the promotion can
just be called again to fix it.

Once a promotion is executed, it makes box.cfg.read_only
attribute be immutable.

The promotion has several API methods:

* box.ctl.promote({timeout = ..., quorum = ...}).
  This function is meant to be called on a slave to demote the
  old master if exists and promote the current instance.

* box.ctl.demote({timeout = ..., quorum = ...}).
  The same a promote, but the procedure is stopped after a master
  is demoted. But note that nevertheless the master still is a
  master, in a read-only mode though.

* box.ctl.promote_info().
  This function shows info about the latest promotion (finished
  or running now - does not matter, just the latest).

* box.ctl.promote_reset().
  This function clears the promotion history so a user would be
  able to re-assign master/slave roles in a cluster manually,
  change box.cfg.read_only etc.

To enable all those API a one should set its box.cfg.read_only
to box.NULL.

Closes #3055

@TarantoolBot document
Title: Document box.ctl.promote()
Subj. For details of the patch see the commit message. For
details of the protocol and API see the RFC:
doc/rfc/3055-box_ctl_promote.md