box: update collation strength option.
At the moment all collations that don't have their "strength"
option set works the same way as ones with this option set to
"identical". It is not efficient, according to ICU Comparison
Levels. This patch updates this option of old collations and set
its default value to "primary" for new ones.

Closes #3573