swim: allow two task types - static and dynamic
Static task is allocated on a static or external memory and
should not be deleted on the scheduler destruction or the task
completion. It works for preallocated tasks like round_step_task,
member.ping/ack_task.

Dynamic tasks are going to appear in a commit, introducing
indirect pings, where each member can send multiple pings and
acks at the same time.