box: rework memtx_tree as an abstract class
The memex tree class was reworked as template using preprocessor
directives and was moved to a separate header file. Now it is
possible to declare a memtex tree for arbitrary leaf structures
that include the basic representation omemtx_tree_key_data and
memtx_tree_data and their comparators.

Needed for #1260